refactor(ui): split cars panel sections
